rename textStyles to fontStyles
