+++ frontend/src/theme/forms/inputStyles.tsx
@@ -1,5 +1,5 @@
 import type { ComponentStyleConfig } from "@chakra-ui/theme";
-import textStyles from "../textStyles";
+import textStyles from "../fontStyles";
 
 const Input: ComponentStyleConfig = {
   sizes: {
